backend/requirements.txt
# =============================================================================
# KIRA Backend — Python Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# --- Data Validation ---
pydantic==2.6.3

# --- Environment ---
python-dotenv==1.0.1

# --- Google Gemini API ---
google-generativeai==0.4.1

# --- HTTP Clients (for Maps/OSM APIs) ---
httpx==0.27.0
aiohttp==3.9.3

# --- Database ---
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1

# --- Utilities ---
Pillow==10.2.0
numpy==1.26.4

# --- Testing ---
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0

# --- Linting ---
ruff==0.3.0
